requirements.txt
asgiref==3.11.1
charset-normalizer==3.4.4
Django==6.0.2
joblib==1.5.3
mysqlclient==2.2.8
numpy==2.4.2
pandas==3.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
requests==2.32.5
scikit-learn==1.8.0
scipy==1.17.
sqlparse==0.5.5
urllib3==2.6.3
